backend/pyproject.toml
[project]
name = "ocr-idp-backend"
version = "0.1.0"
description = "Local OCR IDP backend powered by Qwen2.5-VL and Ollama"
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "pymupdf>=1.24.0",
    "pillow>=10.0.0",
    "requests>=2.32.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
    "pytest-asyncio>=0.23.0",
]
